chore: move native modules to @berty/native-modules and remove dead code
