+++ js/index.ts
@@ -10,10 +10,10 @@ import 'string.fromcodepoint'
 import protobuf from 'protobufjs'
 import { AppRegistry, NativeModules } from 'react-native'
 
-import BridgeLogger from '@berty/go-bridge/logger'
 import { initI18N } from '@berty/i18n'
 import App from '@berty/messenger-app/App'
 import { name as appName } from '@berty/messenger-app/app.json'
+import BridgeLogger from '@berty/native-modules/GoBridge/logger'
 
 protobuf.util.toJSONOptions = { longs: String, enums: Number, json: true }
