+++ js/packages/native-modules/GoBridge/index.ts
@@ -1,0 +1,1 @@
+export { GoBridge } from './GoBridge'
