+++ js/packages/native-modules/PushTokenRequester/PushTokenRequester.native.ts
@@ -1,0 +1,5 @@
+import { NativeModules } from 'react-native'
+
+import { PushTokenRequesterInterface } from './PushTokenRequesterInterface'
+
+export const PushTokenRequester: PushTokenRequesterInterface = NativeModules.PushTokenRequester
