+++ js/packages/native-modules/PushTokenRequester/PushTokenRequester.ts
@@ -1,0 +1,9 @@
+import { PushTokenRequesterInterface } from './PushTokenRequesterInterface'
+
+class NoopPushTokenRequester implements PushTokenRequesterInterface {
+	request(): Promise<string> {
+		return Promise.reject()
+	}
+}
+
+export const PushTokenRequester: PushTokenRequesterInterface = new NoopPushTokenRequester()
