+++ js/packages/native-modules/PushTokenRequester/PushTokenRequesterInterface.ts
@@ -1,0 +1,3 @@
+export interface PushTokenRequesterInterface {
+	request(): Promise<string>
+}
