+++ js/packages/native-modules/PushTokenRequester/index.ts
@@ -1,0 +1,1 @@
+export { PushTokenRequester } from './PushTokenRequester'
