+++ js/packages/utils/accounts/defaultCLIArgs.ts
@@ -1,0 +1,9 @@
+export const defaultCLIArgs: string[] = [
+	'--node.display-name=',
+	'--p2p.swarm-listeners=/ip4/0.0.0.0/tcp/0',
+	'--p2p.high-water=60',
+	'--p2p.low-water=40',
+	'--p2p.webui-listener=:3000',
+	'--log.format=console',
+	'--log.filters=info+:bty*,-*.grpc warn+:*.grpc error+:*',
+]
